[PORT FROM K34]: apic: fixing the mwait hang issue during s3 exit
BZ: 16590

This patch was missed in the pmu driver s3 enabling. This is a very
important patch where we dont disable lapic when we are entering S3.
In cases where there is a mwait failure & scu aborts, pmu interrupt
will not be forwarded to IA as lapic is suspended & system will hang
with bothcpu's in mwait state. hence in this patch we try to have a
seperate lapic_suspend/resume routines specific to our platform.

Original patch was altered by David Cohen in kernel 3.4. This patch
merges the changes done by David and Vishwesh.

Change-Id: I8903d353ea038280d6276f44c25e8b36fb2458bf
Signed-off-by: Vishwesh M Rudramuni <vishwesh.m.rudramuni@intel.com>
Signed-off-by: Sathyanarayanan Kuppuswamy <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-on: http://android.intel.com:8080/128416
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>