Audio: SSP Platform/DAI handle DMA LLI chained mode
BZ: 143599

Instead of programming a new DMA transfer after
each end of them, DMA LLI chained mode is now used.

This should avoid some perf issues: after the end
of a DMA transfer, a tasklet is triggered but proc
can be preempted before the tasklet is executed and
next transfer is programmed.
And the time to program the new transfer is very short:
the duration of the SSP FIFO.
This can lead to some glitches.

Change-Id: I4f274ad722726d88b1cbe8771da74e0e32f481de
Signed-off-by: Sylvain Pichon <sylvainx.pichon@intel.com>
Signed-off-by: Apelete Seketeli <apeletex.seketeli@intel.com>