cpufreq: sfi-cpufreq read back correct values
BZ: 127716

When sfi-cpufreq driver programs a P-State transition, the driver
tries to program a frequency and a voltage (through IA32_PERF_CTL
register, setting BUS_RATIO_SEL and VID_SEL).

The readback of values doesn't give back the exact values set. The
voltage set is not the same as the voltage asked for. This has
triggered previously a patch to ignore hardware read back values,
and use a "cached value".

As the cached value is false, because it doesn't take into account
hardware constraints (such as bus frequency sharing between cores,
frequency shared between hyperthreaded cpus of the same core),
this patch fixes the readback value (ie. the one read back from
/sys/devices/cpu/cpu<N>/cpufreq/cpuinfo_cur_freq).

To work-around the voltage problem, only the bus ratio is used
to find which P-state the CPU is in (as opposed to bus ratio and
voltage couple). This gives an accurate value for all CPUs
actual frequency.

Also, fix some trivial coding style issues.

Change-Id: Idb7ef2dc1227a068b11600881eaf54e252c1477a
Orig-Change-Id: I9f2a6691204ff74c1decc8ee83c1840d542b4f31
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Reviewed-on: http://android.intel.com:8080/92613
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
[Fix trivial style issues]
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>
Reviewed-on: http://android.intel.com:8080/126413
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Tested-by: cactus <cactus@intel.com>