Don't turn on vprog1 from RHB camera drivers on VictoriaBay HW
BZ: 90964

The powering of the cameras is different on RedHookBay and VictoriaBay.
On RHB, vprog1 is at 2.8V and supplies both cameras.
On VB, vprog1 supplies the secondary camera and must not rise
over 1.2V otherwise the secondary camera will be permanently damaged.

Modify the RHB camera driver platform SW to detect whether the SW has
accidentally been flashed on VictoriaBay HW.
If yes, don't turn on vprog1.

Note: The detection mechanism relies on the SPID.
The below patch must be included in order for the VictoriaBay PR1
HW SPID to be available:
http://android.intel.com:8080/#/c/94174

Change-Id: Iaede0095337df5da4bff6e0e23d71033adad681c
Reviewed-on: http://android.intel.com:8080/95132
Reviewed-by: Toivonen, Tuukka <tuukka.toivonen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>