Audio: Fix dma cookie counter
BZ: 89593

In LLI mode if block interrupts are enabled, then the dma cookie is
reset after first block interrupt. So for the second block interrupt
BUG_ON will be triggered and results in panic.

dma cookie is reset is now done after the complete transfer is done or
if the transfer complete is called with TERMINATE command. With
TERMIATE command if the dma is in progress for that descriptor then
cookie reset is called.

Change-Id: I4d67a50ed27e686077481971d41adaddcbc72254
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Reviewed-on: http://android.intel.com:8080/138956
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Tested-by: Koyili, SrikanthX <srikanthx.koyili@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>