REVERTME: ctp, crashtool: add android compilation dependence
Yes, there are lots of garbage stuff Android projects depending on.
Lots of external kernel projects are referring kernel headers directly.

In you are an external kernel project and not a module, I curse you for
breaking my blank_flashfiles compilation! >:)

This patch will be reverted when real component is added to tree.

Change-Id: Iddd96ec135490afed83cb4cd9b9f56e234703a40
Signed-off-by: David Cohen <david.a.cohen@intel.com>