[PORT FROM R42B-STABLE] drm/i915: Watchdog timeout support
BZ: 149537

Adding support for watchdog timeout on individual batch buffers
(sometimes called "media reset"). This feature is supported for
the render and video rings, but not the blitter ring. This
change requires the TDR patch (BZ: 149550)

The watchdog feature can be requested per batch buffer via a new
flag to the exec buffer IOCTL. So far this is only available on
the KMD side and it will require a separate commit to the
perforce tree to add the define for user mode.

The watchdog works by inserting some extra commands into the
ring just prior to the MI_BATCH_BUFFER_START. This starts a
counter in the hardware. If the counter reaches the threshold
value (pre-calculated during driver initialisation) then an
interrupt is generated and this immediately queues the
recovery work handler (same reset function as used by TDR).

Under normal circumstances the counter should not get near
the threshold as extra commands are inserted after the
MI_BATCH_BUFFER_START to cancel the hardware counter. However,
if the hardware hangs whilst executing the batch buffer then
the timer will not be cancelled and the interrupt will fire.

The render and BSD timer interrupts are unmasked during ring
initialisation and left unmasked. The get_irq/put_irq function
had to be modified as it assumed ownership of the ring IMR
registers.

Additional spin locking has been added to
i915_report_and_clear_eir() function as it can be called from
a kernel timer (soft irq context) or from a real interrupt so
it could be re-entrant.

There is a new module parameter (i915_enable_watchdog) which
can globally enable/disable watchdog support.

Added RCS_T, VCS_T, BCS_T and RCS_W, VCS_W, BCS_W to the debugfs
interface to give separate count of hang detections for TDR (_T)
and watchdog (_W). The existing RCS, VCS, BCS total counts and
full GPU reset counts still exist.

12-11-13 Removed duplicate freq calculation in i915_init_watchdog.

14-11-13 Resolved merge conflicts.

Change-Id: I3cae0d96287737a6d512f504e229b9e93537f314
Signed-off-by: ian-lister <ian.lister@intel.com>