Audio: Turn off ext Amp when no playback on IHF
BZ: 113363

As the Virtual device is always on the dapm path for IHF is always
connected when the display is on. This causes the external amp to be
always on.

A virtual spk map is provided which will be always on when the display
is on so as to keep the codec ASRC and DAC on. When actual playback on
IHF is started the dapm switch can be used to turn on/off the external
amplifier.

Also updated the fmt print macros.

Change-Id: I9b9d5cbea4af5813e66b348711f7ea6fc08553ae
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>