NextLTS: Adding Kernel FT support for new kernels
BZ: 118170

This commit is for adding back in support for module stripping
to our NextLTS (3.9) Linux kernel.

As part of the transition from r42-stable to our current android-main
the logic from kernel-build.sh was moved into the AndroidKernel.mk. As
part of that transition calls to modules_install were removed as we had
logic in Makefile.modpost that would accomplish the logic of stripping
debug symbols. In the transition for NextLTS kernel (from kernel 3.4
and kernel 3.9) in our current android-main the code that we were relying
on in the Makefile.modpost file for stripping out debugging symbols was removed.

The changes in this commit replaces the module stripping logic so we no longer
will hit the OSIP header size limit due to module symbol bloat.

Change-Id: I903243dad5ff18d437201fb6a34859dcc81471a0
Signed-off-by: Brian Wood <brian.j.wood@intel.com>