BACKPORT: ufs: don't disable_irq() if the IRQ can be shared among devices
BZ: 179403

When removing the UFS driver, disable_irq() is called and the IRQ is
not enabled again.  Unfortunately, the IRQ is requested with IRQF_SHARED
and it can be shared among several devices.  So disabling the IRQ in
this way is just breaking other devices which are sharing the IRQ.

upstream commit ID: e0f4b7c5acbac65d2f56ed5e1d2d8036f9db936e

Signed-off-by: Akinobu Mita <mita@fixstars.com>
Signed-off-by: Santosh Y <santoshsy@gmail.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Change-Id: I251c4d6c7215d870d5ac8c006d371c7e086470ed
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>