Audio: Fix to get the LPE out of reset during kzalloc failure
BZ: 132116

If the kzalloc failure happens during the codec download, especially
when allocating memory to cache the firmware then LPE is not
got out of the reset state.

The caching of firmware and parsing of firmware binary functionalities,
need not be done having LPE in reset state. This addresses the problem of
not getting LPE out of reset when the memory allocation fails for caching
the firmware. In addition to this, it also reduces the time for LPE in reset
state especially when a PCM playback was ongoing and an offload stream was opened.

Ported from K3.4 BZ: 116607

Orig-Change-Id: Ibb63be519c1fe9b57b36a8e4c4fead8eac4b7914
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/121757
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Griffiths, Richard A <richard.a.griffiths@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
(cherry picked from commit 9e290ce1b32d545bd398a34fd0d835d017458a6d)

Change-Id: I4622adf76e1e6e88e7043ab5ee5f990566362acf
Reviewed-on: http://android.intel.com:8080/132774
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>