[PORT FROM K3.4]mmc: generate the interrupt unexpected
BZ: 79471

When host generated the CMD conflicts interrupts, driver needs a small
delay before clearing it otherwise such interrupts will be generated
again after clearing it. This behavior is not described in SDHC spec,
and suspect it is a silicon issue.

Before confirm whether this is a silicon issue, have a small delay is
still an avaliable software workaround

Orig-Change-Id: Ie3ffa8b4c63af3eada834de879a351ca6bb5a1c0
Change-Id: I019989f72dc22683386a39a6aa2e0a6ffb690643
Signed-off-by: Ning Feiyi <feiyix.ning@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/126370
Reviewed-by: Gao, Yunpeng <yunpeng.gao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>