[PORT FROM R42B-STABLE] platform_sdhci_pci: set ACPI handle back to NULL
BZ: 127480

if  D3cold  is not explicitly allowed, pci will
try to set the device to D3hot. on the current
acpi table there is no PR3 resource defined for
sdchi, only PS3, which means that only
D3cold will be valid.

If D3hot was targeted and fails, the code defaults
to the next power state which would be D0 preventing
the platform from correctly entering S3.

Set the ACPI handle to null after reading the
gpio data, so that pci does not try to do ACPI PM.

Change-Id: I747bf50d52188ec3c02d01938986e0ccf639e4af
Orig-Change-Id: I747bf50d52188ec3c02d01938986e0ccf639e4af
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/137523
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>