[PORT FROM R43-STABLE]audio: Fix to turn off the MIC BIAS when the headphone is inserted
BZ: 125263

This patch turns off the mic bias when the headphone is inserted.
MIC BIAS needs to be "on", only when the headset is inserted, inorder
to detect the button press events. Wheres, this was "on" for headphone
as well. By doing this, we can save the Mic bias power during
headphone insertion.

Orig-Change-Id: I9dcc34319603684ea5077d6a94ddec8c30d58ae0

Change-Id: I8c74dd01a8b3dc27265182842fe5910f138cdb84
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Signed-off-by: Zhou, Ting Z <ting.z.zhou@intel.com>
Signed-off-by: sheetal <sheetalx.garg@intel.com>