[AUDIO] Set the FSRT bit for SSP0 transfer protocol
BZ: 132116

The frame signal defines the boundaries of a basic data unit, comprised
of multiple serial bits.

The FSRT bit (frame sync relative timing) is part of the configuration
of the programmable serial protocol used for SSP transfers.

When set, the frame signal corresponding to the next sample will
be asserted during the transmission of the LSB of the current sample.
When cleared, this bit causes the frame to be determined in another way
that is not aligned with the SSP transfer protocol that is programmed
between slave and master, which leads to noise in the Rx direction
on data received on SSP0 port.

This patch sets the FSRT bit.

Ported from K3.4 BZ:122522

Orig-Change-Id: I8b745ed4b2b988a9c379310ae6d28b536813e7d7
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>
Reviewed-on: http://android.intel.com:8080/123141
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Pichon, SylvainX <sylvainx.pichon@intel.com>
Reviewed-by: Cheverneuil, FredericX <fredericx.cheverneuil@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
(cherry picked from commit 625b1189495153b065c7708bf48779fe5ef62e0a)

Change-Id: I31e8d6ddc886ec367c7871a97d79d54314c5e53e
Reviewed-on: http://android.intel.com:8080/132796
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>