Audio: Added support for mutiple substream in ASOC
BZ: 132116

Currently in ASOC framework dai supports one playback and
one capture stream.

This patch adds support for multiple substream for playback and
capture by adding a counter in dai structure to specify the no of
substreams required. This counter will be passed to snd_pcm_new
which created the required subdevices for device.

For K3.4:
	BZ: 85734.
	Reviewed-on: http://android.intel.com:8080/91968

Change-Id: Ia6f764f8c0bb652f6c54216bd273a84509b8e5e0
Orig-Change-Id: I5b35406e44bd67d9248b34f66dabf41671d25054
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Praveen Diwakar <praveen.diwakar@intel.com>
Reviewed-on: http://android.intel.com:8080/123305
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>