[PORT FROM K3.4][REVERTME] sdhci-pci: power up Merrifield SDHCI hosts
BZ: 94574

After enabling D0ix on sdio0/1/emmc0 and rebooting the platform

The issue is, emmc0 is in d0i3 during reboot and there is a access
without bringing the device to d0i0.

Fixed this by implementing shutdown handler and disabling runtime_pm
during shutdown, this is similar to other sdhci driver implementations.

However this solution is a workaround since during shutdown emmc driver
is not expecting access, and since there is access need to find who
is accessing and fix that issue.

Till that time we have this solution to unblock S0ix enabling.

This patch also includes:
 bd46f3f8 [REVERTME] sdhci-pci: implement shutdown callback

Org-Change-Id: If85dc9829a878e9129f3aa5055b6d74469043733
Change-Id: I68886dca52cd2cf7e7fc6c20dfb22fef930fd3b5
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/126367
Reviewed-by: Gao, Yunpeng <yunpeng.gao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>