[PORT FROM K3.4]mmc: sd: workaround VSDIO off issue on BYT
BZ: 124063

Workaround for vallyview sighting 4684219.

On BYT, VSDIO and VDD will be turned off when LPSS and SCC are in D3
which will power off the SD card when SD host is in D0i3.

This is not what we want. We expect the SD card is still powered on when
SD host is in D0i3.

Here is a workaround solution on BYT:
Change 1P8_EN and PWR_EN pins to GPIO mode instead of working at native
function. In this case, the power control register enable bit and host
control 2 1.8v signaling bit cannot control the VDD and VSDIO, and driver
needs to write GPIO to control them

Change-Id: I99767a79aa32708df5d7bf9a8f7297d205b7f56b
Orig-Change-Id: I76b7a4ad53617a8c2436f000e60dff0d1e957f03
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/127364
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Gao, Yunpeng <yunpeng.gao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>