n_gsm: Rework of close procedure
BZ: 74701

This patch is to allow the mux to close properly all DLCI as gsmld_close cannot :

When gsmld_close is called, tty can't be used anymore !
Here is the 3 possible cases :
      - Close of the tty on which the mux is associated
        tty->ops->close is called before tty_disc_release
      - Hangup of the tty on which the mux is assocated
        A modem/link problem occurs and modem is not reachable
      - Replacing LD by another one
        tty->receive_room is set to 0, no modem answer can be
        received

So, no need to send anything to modem, ether it will not
reach the modem, the modem can't accept it, or
we will not receive modem answer and we will timeout (7.5s).

To close modem mux, a new ioctl (GSMIOC_DEMUX) have been
added. This need to be called before the close ttyIFX0.
This ioctl close all DLCI sending DISC and waiting for UA.
This will also avoir 30s timeout from tty waiting incomming
command (DISC) to be sent to HSI link.

Now, gsmld_close only removes structures and doen't try to communicate
with the modem.

Change-Id: I227fe8874bfbfc56059d9a5fc1d6ec5e36901635
Signed-off-by: Vincent Pillet <vincentx.pillet@intel.com>
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>
Reviewed-on: http://android.intel.com:8080/128580
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>