[PORT FROM K3.4][Telephony/MCD] Handle new SFI table entry
BZ: 99672

This update shall not break platforms on which the SFI entry is not yet
created. Previous handling is keeped during the transition.

This patch includes:
- new registration function. This function allow sfi table handler to register
a structure that will be used by MCD.
- New platform data structure creation function. Instead of static structure,
MCD will create dynamically its own configuration, based on data given by sfi
table.

The aim is to reuse most of the code originally created.

Change-Id: I303d847cbf5bc0c005f02efc2ddd078269bb57a1
orig-Change-Id: I97ff8cb71c89e6baf8dcc1e12ad9fe3a9cd3bb74
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>
Reviewed-on: http://android.intel.com:8080/125331
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>