[PORT FROM K34]dwc3:gadget: set maxpacket to default value when ep is disabled:
BZ: 108778

Need to reset the maxpacket value of endpoint when it is disabled.
As gadget driver will config the maxpacket value of endpoint, but
the value is not reset to default value after ep released by gadget
driver. This will cause wrong maxpacket issue if the same endpoint
assigned to another gadget function driver.

Change-Id: I4480ebc67ed47bdc21f3d4d6ba7442770ab65c5c
Orig-Change-Id: I09ai6e2da0317f8dae19a1ce966077beed855c869
Signed-off-by: Wu, Hao <hao.wu@intel.com>
Signed-off-by: Jiebing Li <jiebing.li@intel.com>
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/129760
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>