drm/i915: cover ring access with rpm get/put
Covering the ring activity is a little tricky. To make sure Gfx is
in D0 while there are still commands pending on ring following is
done.
- All the ioctls are already covered with get/put this makes sure
at the time of scheduling commands on GPU, Gfx is in D0
- Once command scheduling is done, we call add request to track ring
activity.
- We do get_ring (which calls runtime_noresume if this is first request).
in the add_request function. This will only increment the usage_counte
as we are already in D0, we don't need to call get.
- put is called only when request_list becomes empty. i.e gpu is idle
and there are no pending commands on the rings

Introduced a new function i915_add_request_wo_flush, which is similar
to i915_add_request without doing a flush. We need this to track
ring activities which are not covered with i915_add_request.

Note: Make sure we don't do multiple add_request in same ioctl/callback
only one in the end is enough.

Change-Id: I30c100254829f2e478e5b0e60f018a884359e969
Signed-off-by: naresh kumar kachhi <naresh.kumar.kachhi@intel.com>