audio: Moving period elapsed reply handling to the interrupt handler
BZ: 142257

This patch moves the period elapsed reply handling from the threaded irq
to the main interrupt handler. The period elapsed interrupts from firmware
are received back to back and handling them in the top-half ensures that
the period elapsed message from firmware is not missed. Having them in
the threaded interrupt handler might lead to the miss of interrupt, since the
threaded irq runs in the process context.

Change-Id: I638b458fd90bd51259d9f0cbf0b731edbe514af4
Orig-Change-Id: Iadc2271dd0a52a988533c51bf3aea2f7df464872
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Signed-off-by: Hardik T Shah <hardik.t.shah@intel.com>
Reviewed-on: http://android.intel.com:8080/135484
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Tested-by: Koyili, SrikanthX <srikanthx.koyili@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>