drm/i915: Added the 128/256 cursor support
BZ:97888

Tested the 64-bit cursor.Not tested the 128 & 256 cursor mode.
Couldn't get any input layers from the Surfaceflinger in that range.

Even in the 64-bit mode getting some buffer corruption in the end.
The reason is the input cursor resolution is 39x30 but the expected
size is 64x30.

The suported resolutions are 64x64, 128x128 and 256x256.

Change-Id: Icbb2f7e12bf9532fa51bb7059af478c34118dbfd
Signed-off-by: Pallavi <pallavi.g@intel.com>
Reviewed-on: http://android.intel.com:8080/100209
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>