BTY GFX K3.10 : Make config item available for Kernel patches
There are about 809 patches backported from upstream kernel tree
that needs to be applied to the kernel-next branch for MCG
in order to get Baytrail GFX UI to come up in Kernel 3.10.

One of them is a Kconfig items. This patches sets the default
config so that kernel building would go smoothly after the
patches are applied

Change-Id: I1a2a1d3325cf31fb506c8ff5d18f3fbd07369608
Signed-off-by: Vinil Cheeramvelil <vinil.cheeramvelil@intel.com>