gpio-keys: Remove home btn for BYT-T
BZ: 146451

Android OS has provided virtual home key, which is the
recommended way of using home button. We do not need to
support home button for byt-t boards, and there isn't
such hardware key on the front of the byt-t ffrd boards.
We remove the home button key from the key list.

Change-Id: I990ac019233e8f7d990cf083c1d90fa7ad82e736
Signed-off-by: Li, Ning <ning.li@intel.com>