PTI driver: add ability to write segmented messages to PTI
BZ:137218

Applications occasionally need to write a trace message involving multiple
memory areas. As a consequence, the PTI driver shall provide the possibility to
write a trace message in several write operations. An Enf Of Message (EOM) indication
is introduced. The DTx signal shall only be sent if EOM is true. This allows to
write a trace message involving multiple memory areas without forcing extra copy
operations on the data.

Change-Id: I4fe2595d7b727b1d854e2e630bcb253e35f3a95b
Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>
Reviewed-on: http://android.intel.com:8080/130942
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>