[PORT FROM R42B-STABLE]Thermal: Fix interrupt handling in thermal driver
BZ: 132483

The interrupt handler in the thermal driver looks
for a single event upon every hardware event; i.e
it checks the first bit set in the status register
and handles that event and exits the handler.
Sometimes more than one event happens for a
single hardware interrupt.
This patch makes the interrupt handling code to look
for multiple events whenever an interrupt occurs.

Change-Id: I880b60b42226fde1a5962896d8bc40607d1cca68
Signed-off-by: Sumeet Pawnikar <sumeet.r.pawnikar@intel.com>
Reviewed-on: http://android.intel.com:8080/131588
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>