PTI: implement D8/D16 STP format support for Tangier and Carbon Canyon
BZ: 138224

Support D8, D16, DT8 and DT16 STP formats to avoid extra zero-padding of
trace messages with a length that is not a multiple of 4 bytes.

Currently, the PTI driver always write a trace message in chunks of 4 bytes
into the PTI memory apertures. pti_write_to_aperture() function shall be updated
so that it can also write chunks of 1 or 2 bytes into the aperture.

Note that this feature is not supported on all SOCs. Intel Tangier SOC does support this.
Intel CRC SOC (Carbon Canyon) will also support this.

Change-Id: I1a5287d7860cd377fb4b157b27b7958d9d6c773e
Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>
Reviewed-on: http://android.intel.com:8080/131807
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>