Audio: Fix for incorrect mask values of SOC_VALUE_ENUM_SINGLE in cs42l73
BZ: 139005

The mask and value computation in k3.4 for the SOC_VALUE_ENUM_SINGLE
was based on the xmax argument, whereas the same is computed based on
xmask in k3.10. Because of which there was an incorrect value written
to the SPK Loopback Mux which resulted in a noisy audio during
simultaneous playback and capture on cs42l73 codec. This patch
adds the correct mask value in CS42L73 codec driver wherever the
SOC_VALUE_ENUM_SINGLE is used.

Change-Id: I7a1c2ccc78c09338466016bc015b3f6ae93fa238
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/132195
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Shah, Hardik T <hardik.t.shah@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>