drm/i915: Increase the pin count size in GEM objects
VLV HWC is hitting a limit on the number of bits in the pin count.
For the Widi case, we are reaching the maxlimit of 15 with 4 bits,
so time being using 5 bits to provide additional headroom.
Todo, understand that why 4 bits are proving insufficient.
One reason could be due to late scheduling of unpin work items.
The un-pin(decrement of pin count variable) of the frame buffer
GEM object is performed from a work item queued in the workqueue.
And Widi is a stressful scenario where lot of activities take
place, like there is a simultaneous decode/composition/encode
& also streaming is done, so the un-pin work items scheduling
could get delayed. Other reason could be a genuine requirement
of a higher pin count in Widi scenario.

Change-Id: I75e6063a0054bfc950268d9b45bf42e8c040495d
Signed-off-by: akashgoe <akash.goel@intel.com>