[REVERTME]PCI: Ensure platform pci operation is called for D0
BZ: 133997

This allows the platform PCI code to be called for D0.
Without this, there seems the devices are powered down
by f/w and the pci_enable_device skips the platform power
transitions thinking that calling it is unnecessary as the
current_state is not updated.

This is a temporary w/a and the the right solution to be
addressed in the BZ: 140519

Change-Id: Ia6a6942a403d382df0a295202ec3ac020c455a39
Signed-off-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>
Reviewed-on: http://android.intel.com:8080/132855
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>