audio: Change the event format for VTSV
BZ: 139094

The driver was sending the VTSV result data from the FW to userspace
as a series of strings. Each string contained the decimal value of
the corresponding data byte.This has two issues.
1.The Java layer that receives the data needs the data in
"key=value" format.
2.The result data fields are in 16 bit. Sending them as bytes requires
the user space to assemble them into 16 bit words

This patch sends the result data as "DATAn=d", where "n" is the
index of the 16bit data word and "d" is the decimal value of the
16 bit data word.

This patch also adds a key-value pair to specify the event type
in the uevent so that userspace can poll on this.

Change-Id: Ibab5f156a62f102772f309d55e5935baa2c8c305
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>