[PORT FROM K3.4][FFRD8:EM] Ported SMB347 and MAX17042 drivers from K3.4 to K3.10, config changes
BZ: 141437

Description:
SMB347 and MAX17042 drivers are not enabled in K3.10 for BYT-FFRD8
platform.

Solution:
With ACPI entry added in IFWI, this patch will enable the SMB347 and
MAX17042 drivers and will be probed only in the target which has
SMB347 and MAX17042 chips.

Category: device enablement

Domain:EM-BYT_FFRD8

Origin: Internal

Change-Id: Ie4a5feba672ebd65accde104ed8d25d945e3722a
Signed-off-by: Venkataramana Kotakonda <venkataramana.kotakonda@intel.com>
Reviewed-on: http://android.intel.com:8080/134760
Reviewed-by: Kannappan, R <r.kannappan@intel.com>
Reviewed-by: Pandey, Raj <raj.pandey@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>