extcon: arizona: Get pdata from arizona structure not device
In the case of a device tree system there will be no pdata attached to
the device, causing us to deference a NULL pointer. Better to take the
pdata from the Arizona structure as this will always exist and we know
will have been populated since it is populated by the MFD device which
binds in the extcon driver.

Change-Id: I204cff353a38149312238787c99f17cfa78196bb
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>