audio:Baytrail: Delay reporting of headphone
BZ: 120635

The headset detection algorithm implements retries to detect accessory.
If the accessory is not HS (i.e HP or no accessory) retries are done.
During slow inserion of HS, first a HP would be detected and then HS.
If we report HP and then HS, the audio pauses due to accessory change
handling in android middleware.

Inorder to avoid this, if HP is detected, it is not reported
immediately. HP is reported if the accessory is still HP on the last
retry.

While the audio pause during slow insertion of HS would be avoided,
this change would increase the delay in reporting HP since it would be
done at the end of all retries. This is a known and accepted tradeoff.

Also changed the number of retries from 3 to 6 and the retry interval
from 300ms to 100ms for faster response.

Change-Id: I8cb4a85c00f2de0f5ffff808fbf93f8b7ce8a68f
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>
Reviewed-on: http://android.intel.com:8080/136898
Reviewed-by: K, Mythri P <mythri.p.k@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Shah, Hardik T <hardik.t.shah@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Tested-by: Koyili, SrikanthX <srikanthx.koyili@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>