[HACK]apic: Do not disable apic during S3
BZ: 142329

In cases where mwait fails or scu aborts and if apic is
suspended the system will hang as interrupts will not be
forwarded. There is a similar patch for intel-mid platforms
and since BYT is no longer enables intel-mid, we need to
add a similar fix for Baytrial as well

Created a new BZ:142523 wherein the more generic solution
will be done so that it can be easily upstreamable.

Change-Id: I9801a650b76db455d6ce539ff9b1c7cee2af57a5
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>
Reviewed-on: http://android.intel.com:8080/135585
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Tested-by: Gongora, Enrique <enrique.gongora@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>