[PORT FROM K3.4]telephony: Scalability V2
BZ: 139961

This patch is the first one of a bunch of 3 ones to implement
the telephony scalability V2. This patch creates a new SYSFS
indicating which configuration is running. This information is
determined thanks to CPU type and modem type.

The second patch will prepare the SFI/ACPI data change.
Currently, those tables contain the CPU name and the modem name.
It will be replaced by the configuration name only. For
compatibility reasons, the 3 SYSFS (cpu_name, modem_name and
config_name) need to be correctly populated regardless of
SFI/ACPI tables content.

Third patch will remove this compatibility. Userland apps need to
be updated first.

Change-Id: I449ba9619ce9de8fc849b82f78a3f54286309ef2
Orig-Change-Id: I7052b23245c870eb20211cd0f54e235e9a01041a
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/136289
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>