audio: Removing the workqueues for process_msg and process_reply
BZ: 140432

This patch removes the single threaded workqueues of process_msg
and process_reply of clovertrail and merrifield. All the replies from
firmware are added to list in the interrupt handler and the clearing
of interrupt mask registers are also done in the interrupt handler
itself. This list is iterated and the messages are processed in the
threaded handler. This way, there is a common threaded handler.
In addition to this, the period elapsed reply handling has been
moved to the respective process reply handlers.

Change-Id: I1cb777b41406ae7074d0f1a43a2ba6f54ebd5e81
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/136507
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Koyili, SrikanthX <srikanthx.koyili@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>