Put back the Intel modem NVRAM driver for kernel 3.10
BZ: 142088

The driver which implements the modem NVRAM interface.
- size file : read it to get NVRAM size (decimal)
- dump file : read or write to communicate with this NVRAM
- clear file : any write access will clear (zeroing) the NVRAM

Driver is started or not, depending on the corresponding flag
set in the defconfig files.

In intel_modem_nvram driver check intel_scu_ipc_get_nvram_addr()
return value for merrifield. If return value is different than 0
modem data caching in OSNIB is supported (starting on B0 silicon
revision). In this case store data to OSNIB and issue merrifield
IPC for data copy to DCCM SCU memory.

Change-Id: Ie5fd1f5bf046e1532e50b2ebc39bf9116001d05e
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/136357
Reviewed-by: Touzeau, GabrielX <gabrielx.touzeau@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Auger, Florent <florent.auger@intel.com>
Reviewed-by: Romieu, Benoit <benoit.romieu@intel.com>
Reviewed-by: Griffiths, Richard A <richard.a.griffiths@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>