[REVERTME] Audio: Fix no audio at uplink for first CSV MO call
BZ: 142816

In K3.4 AIF1CLK/fs ratio selection is independent of input clock
frequency for WM8958 codec. In K3.10 AIF1CLK/fs ratio selection
is function of input frequency. Provided temporary workaround
to have AIFCLD1/fs ratio to be independent of input clock  frequency.

Input clock frequency to codec must be 19.2MHz for voice and audio.
Currently its 32KHz for voice and 19.2MHz audio. DPCM will fix
same input frequency for Audio as well as voice. Once it's done
this patch will be reverted back.

Input clock frequency to codec is switched from 32KHz to 19.2MHz
as a part of hw_params. Which gets called only for PCM and
compressed DAIs. Switching to 32KHz from 19.2MHz is done as a
part of DAPM OFF sequence. Which gets called for PCM, compressed
and voice calls as well. So for voice call input clock frequency
is not deterministic.

Change-Id: I2f3227e8a135e45f7f572f4b6239e61b04a55226
Signed-off-by: Hardik T Shah <hardik.t.shah@intel.com>
Reviewed-on: http://android.intel.com:8080/136557
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: K, Mythri P <mythri.p.k@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Tested-by: Koyili, SrikanthX <srikanthx.koyili@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>