SEP54: add support for hardware key based CBC AES 256 algorithm
BZ: 136675

This patch adds support for a hardware key based CBC AES 256 async
crypto algorithm. It is available by requesting it via
"cbchk(aes)". This algorithm expects a key id instead of a full key
and it is using the key id to match a key stored in hardware that it
is going to be used for the actual encryption and decryption
operations.

Although it is implemented as a async cipher it is in fact sync as the
current SEPAPP API does not provide async calls. The hardware disk
encryption implements such an API so we could port it from the disk
encryption branch if needed. However, the sole user of this algorithm
is eCryptFS which uses the crypto APIs solely in a synchronous way.

Category: feature differentiation
Domain: AOSP.Extend-Arkham
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: Ifecc47302d43577c72526dd165cd3be877aa490d
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>