[PORT FROM K3.9] drm/i915: Porting of TDR v1 from VLV/HSW to BYT.
BZ: 143942

HSW version: Idc7921fd5decb855ac741706ca51af61a6055d51

Original Authors: Siluvery, Arun <arun.siluvery@intel.com>
		  Lister, Ian <ian.lister@intel.com>

Timeout-Detection-Recovery (TDR) provides per-engine hang detection
and recovery.  If an engine hangs then the TDR will attempt to reset
the engine and advance the command streamer to the next instruction in
the ring. If it was in the middle of processing a batch buffer then
control returns to the instruction following the batch buffer start
command. This provides a less intrusive recovery mechanism than global
reset because it only impacts the process which caused the hang.
uevents are sent so that user mode can detect that something has gone
wrong and take action if required.

TDR updated to handle setting of reset stats when TDR occurs. The
request list is scanned to detect which request caused the hang and
the context stats are updated.

Issues:
If page flips are triggered by submissions to the blitter ring
then a case has been observed where the page flip requests are
continually submitted to the blitter even when the ring has hung.
This prevents the TDR from firing as the timer is reset each time
new work is added to the ring. One solution to this is to only
reschedule the timer if it is not already scheduled.
This problem was not seen on VLV 3.4. It is no longer a problem now
that we are using MMIO register access to trigger the page flips.

Testing was done on BYT devkit using unit tests developed for TDR.
These tests hang individual rings by submitting bad batch buffers,
which will be detected by the TDR logic and it will attempt to recover.
It also tests the cases where the batch buffers are dependent.

NOTE: Whilst resolving a merge conflict, a few lines from another
      patch were lost. I have since reinstated them but they will
      show up with incorrect ownership under "git blame".
      This applies to lines 2502-2513 in i915_debugfs.c.
      The original author was Sagar Kamble.

Signed-off-by: Tomas Elf <tomas.elf@intel.com>
Change-Id: I14508afe43af3d7a71308b4ed89e1797d8ebae0d
Reviewed-on: http://android.intel.com:8080/137353
Reviewed-by: Lister, Ian <ian.lister@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: Jin, AlbertX <albertx.jin@intel.com>
Tested-by: Jin, AlbertX <albertx.jin@intel.com>