[K3.10] Fix stack ovevflow due to the depth of rpmsg call back
BZ: 142651

Currently, during rpmsg initalization, each ns_info is passed and processed
recursively. In other words, name service info for the next service is handled
in a function called from the same function call that just processed
the info for the previous one.

This makes the stack memory size for the current instance increase proportionally
to # of name serivices in the list. This may cause stack overflow especially
when there are more than certain number of name services.

To address this issue, this patch changes the code (call back function)
to push ns_info for all existing (pre-defined) services in the receive buffers
at once to make them processed in one rx function call.

With this way, stack depth needed for that rpmsg registeration is limited
no matter how many services we have in future.

Change-Id: I1bf12582e4408239042bf752aa71fa5640592f86
Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>
Reviewed-on: http://android.intel.com:8080/137020
Reviewed-by: Li, Ning <ning.li@intel.com>
Reviewed-by: Renganathan, Prabu <prabu.renganathan@intel.com>
Reviewed-by: Gao, Bin <bin.gao@intel.com>
Reviewed-by: Griffiths, Richard A <richard.a.griffiths@intel.com>
Reviewed-by: Yang, Bin <bin.yang@intel.com>
Reviewed-by: Chen, Jason CJ <jason.cj.chen@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>