[PORT FROM K3.4]USB IPV6 Mac address causes wrong router IPV6 configuration
BZ: 101677

When USB CDC-NCM driver is initialized, a MAC address is defines for this Ethernet interface.
In IPV6 stack, any Ethernet interface creates automatically a link/local address based on MAC.
Even if UE receives a new link/local from Network, IPV6 stack generates or uses automatically
the MAC address link/local insetad of the link/local coming from Network.
This could leads to wrong IPV6 Address assignement and failure in some test cases,
And it could confuse the network because several UE will do a router sollicitation
using same link/local (because the MAC address doesn't change for this cdc interface.

The fix consist to either reset the auto_linklocal (but not supported in theis kernel) or
remove or undefined the MAC address. To do so, we define the MAC length address to 1 (like
many other driver does). of course we do this change only for our device.

This patch contains :
- Klocwork: Static analysis critical error at cdc_ncm.c:633

Change-Id: I07de08a62a27ae077d8b0185d6d4b4a18c70f2e3
Signed-off-by: feng wang <feng.a.wang@intel.com>
orig-Change-Id: I7447c1bb648397737bd42a07fef425c9cbb73390
Signed-off-by: Muratory Pierre <pierre.muratory@intel.com>
Reviewed-on: http://android.intel.com:8080/137128
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Tested-by: Wang, Ran <ran.wang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>