[PORT FROM K3.4][BYT-M/D:EM] Ported EC and battery drivers from K3.4 to K3.10, config changes
BZ: 135437

Description:
EC and battery drivers are missing on K3.10 and need to port to kernel-next(K3.10) in main line.

Solution:
EC and battery drivers are ported from K3.4 to K3.10. Also ACPI is enabled for driver so that
the driver can be loaded obly in the platforms which has EC.

Category: device enablement

Domain:EM-BYT-M/D

Origin: Internal

Upstream-Candidate: no,not-aosp

Change-Id: I84cff125db2364b869e90685bfd5456244d625ba
Signed-off-by: Venkataramana Kotakonda <venkataramana.kotakonda@intel.com>
Reviewed-on: http://android.intel.com:8080/137208
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Li, Ning <ning.li@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>